refactor(03-01): extract stats.rs into stats/ directory module
- Split 1,864-line stats.rs into 6 focused sub-modules: mod.rs, session.rs, aggregation.rs, persistence.rs, cache.rs, tests.rs
- Replace 3 raw rusqlite::Connection::open calls with SqliteDatabase methods (get_session_start_time, get_session_active_time)
- Add allow(unused_imports) on lib-only re-exports to fix binary-target clippy warnings
- Fix pre-existing database module unused import warnings (SessionWithModel, MaintenanceResult)
- All 246 lib tests pass, all integration tests pass, zero clippy warnings
